fix: refactor class hierarchy
