+++ projects/bin/train_model.py
@@ -1,8 +1,9 @@
 #!/usr/bin/env python
 import sys
 import argparse
-from semantic_selector.model.one_to_one import LsiModel, NNFullyConnectedModel
-from semantic_selector.adapter.one_to_one import MySQLTrainingAdapter
+from semantic_selector.estimator.lsi import LsiEstimator
+from semantic_selector.estimator.fnn_simple import FNNSimpleEstimator
+from semantic_selector.adapter.training import MySQLTrainingAdapter
 
 
 def main():
@@ -27,25 +28,26 @@ def main():
 
     model_name = args.model_name
     print("model type: %s" % (model_name))
-    if model_name == "nn_fc":
-        model = NNFullyConnectedModel()
+    if model_name == "fnn":
+        model = FNNSimpleEstimator()
         options = {
             'threashold': args.threashold,
             'ratio_test': args.ratio_test,
             'seed': args.seed,
         }
         adapter = MySQLTrainingAdapter(options)
-        model.train(adapter, args.epochs)
-        model.save()
+        model.set_adapter(adapter)
+        model.train({'epochs': args.epochs})
     elif model_name == "lsi":
-        model = LsiModel()
+        model = LsiEstimator()
         options = {
             'threashold': args.threashold,
             'ratio_test': args.ratio_test,
             'seed': args.seed,
         }
         adapter = MySQLTrainingAdapter(options)
-        model.train(adapter)
+        model.set_adapter(adapter)
+        model.train()
     else:
         print("model %s unknown" % (model_name))
         sys.exit(1)
